JWPlayer remove() error
fixed. Also added a blank audio file to be used on pages with no audio;
JWP cannot exist without a file and can’t be removed if it doesn’t
exist. using the dummy file - the player always exists and is
re-initialized to set a real file or the blank file.